chore: reusing dto
