+++ openapi/p4pa-pu-bff.openapi.yaml
@@ -71,7 +71,7 @@ paths:
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/AccessTokenDTO'
+                type: 'AccessToken'
         '400':
           description: Invalid request
         '401':
@@ -288,7 +288,7 @@ paths:
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/DebtPositionTypeDTO'
+                type: 'DebtPositionType'
         '404':
           description: Not Found - Debt position type not found
         '401':
@@ -478,19 +478,6 @@ components:
         description:
           type: string
           description: Description of the error
-    AccessTokenDTO:
-      type: object
-      properties:
-        accessToken:
-          type: string
-          description: The access token
-        tokenType:
-          type: string
-          example: Bearer
-          description: The type of the token
-        expiresIn:
-          type: integer
-          description: Time in seconds until the token expires
     PagedDebtPositionTypeWithCount:
       type: object
       required:
@@ -538,60 +525,3 @@ components:
         activeOrganizations:
           type: integer
           format: int32
-    DebtPositionTypeDTO:
-      type: object
-      required:
-        - debtPositionTypeId
-        - brokerId
-        - code
-        - description
-        - orgType
-        - macroArea
-        - serviceType
-        - collectingReason
-        - taxonomyCode
-        - flagAnonymousFiscalCode
-        - flagMandatoryDueDate
-        - flagNotifyIo
-      properties:
-        debtPositionTypeId:
-          type: integer
-          format: int64
-          description: Unique identifier for the debt position type
-        brokerId:
-          type: integer
-          format: int64
-          description: Identifier for the broker associated with this debt position type
-        code:
-          type: string
-          description: Code representing the debt position type
-        description:
-          type: string
-          description: Description of the debt position type
-        orgType:
-          type: string
-          description: Organization type associated with this debt position type
-        macroArea:
-          type: string
-          description: Macro area related to the debt position type
-        serviceType:
-          type: string
-          description: Service type associated with the debt position type
-        collectingReason:
-          type: string
-          description: Reason for collecting this debt position
-        taxonomyCode:
-          type: string
-          description: Taxonomy code for this debt position type
-        flagAnonymousFiscalCode:
-          type: boolean
-          description: Indicates if an anonymous fiscal code is allowed
-        flagMandatoryDueDate:
-          type: boolean
-          description: Indicates if a due date is mandatory
-        flagNotifyIo:
-          type: boolean
-          description: Indicates if notifications to IO are required
-        ioTemplateMessage:
-          type: string
-          description: Template message for IO notifications
