+++ src/main/java/it/gov/pagopa/pu/bff/controller/AuthenticationController.java
@@ -1,8 +1,8 @@
 package it.gov.pagopa.pu.bff.controller;
 
 import it.gov.pagopa.pu.bff.controller.generated.AuthenticationApi;
-import it.gov.pagopa.pu.bff.dto.generated.AccessTokenDTO;
 import it.gov.pagopa.pu.bff.service.AuthorizationService;
+import it.gov.pagopa.pu.p4paauth.dto.generated.AccessToken;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +19,7 @@ public class AuthenticationController implements AuthenticationApi {
   }
 
   @Override
-  public ResponseEntity<AccessTokenDTO> postToken(String idToken) {
+  public ResponseEntity<AccessToken> postToken(String idToken) {
     log.info("User requested postToken()");
 
     return new ResponseEntity<>(authorizationService.postToken(idToken), HttpStatus.OK);
