+++ src/main/java/it/gov/pagopa/pu/bff/mapper/taxonomy/TaxonomyServiceTypeCodeMapper.java
@@ -1,14 +1,13 @@
-package it.gov.pagopa.pu.bff.mapper;
+package it.gov.pagopa.pu.bff.mapper.taxonomy;
 
-import it.gov.pagopa.pu.bff.dto.generated.TaxonomyCodeDTO;
+import it.gov.pagopa.pu.bff.dto.generated.TaxonomyServiceTypeCodeDTO;
 import org.springframework.stereotype.Component;
 
 @Component
-public class TaxonomyCodeMapper {
-
-  public TaxonomyCodeDTO map(
-    it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyCodeDTO input) {
-    TaxonomyCodeDTO dto = new TaxonomyCodeDTO();
+public class TaxonomyServiceTypeCodeMapper {
+  public TaxonomyServiceTypeCodeDTO map(
+    it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyServiceTypeCodeDTO input) {
+    TaxonomyServiceTypeCodeDTO dto = new TaxonomyServiceTypeCodeDTO();
     dto.setOrganizationTypeDescription(input.getOrganizationType()+". "+input.getOrganizationTypeDescription());
     dto.setOrganizationType(input.getOrganizationType());
     dto.setMacroAreaCode(input.getMacroAreaCode());
@@ -17,8 +16,6 @@ public class TaxonomyCodeMapper {
     dto.setServiceTypeCode(input.getServiceTypeCode());
     dto.setServiceType(input.getServiceTypeCode()+". "+input.getServiceType());
     dto.setServiceTypeDescription(input.getServiceTypeDescription());
-    dto.setCollectionReason(input.getCollectionReason());
-    dto.setTaxonomyCode(input.getTaxonomyCode());
     return dto;
   }
