+++ src/main/java/it/gov/pagopa/pu/bff/service/AuthorizationService.java
@@ -1,8 +1,7 @@
 package it.gov.pagopa.pu.bff.service;
 
 import it.gov.pagopa.pu.bff.connector.auth.client.AuthnClient;
-import it.gov.pagopa.pu.bff.dto.generated.AccessTokenDTO;
-import it.gov.pagopa.pu.bff.mapper.AccessTokenDTOMapper;
+import it.gov.pagopa.pu.p4paauth.dto.generated.AccessToken;
 import it.gov.pagopa.pu.p4paauth.dto.generated.UserInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,7 +13,6 @@ import org.springframework.util.CollectionUtils;
 @Slf4j
 public class AuthorizationService {
 
-  private final AccessTokenDTOMapper accessTokenDTOMapper;
   private final AuthnClient authClientImpl;
   private final String subjectIssuer;
   public static final String CLIENT_ID = "piattaforma-unitaria";
@@ -24,10 +22,8 @@ public class AuthorizationService {
   public static final String ROLE_ADMIN = "ROLE_ADMIN";
 
   public AuthorizationService(@Value("${rest.auth.token-exchange-issuer}") String subjectIssuer,
-                              AccessTokenDTOMapper accessTokenDTOMapper,
                               AuthnClient authClientImpl) {
     this.subjectIssuer = subjectIssuer;
-    this.accessTokenDTOMapper = accessTokenDTOMapper;
     this.authClientImpl = authClientImpl;
   }
 
@@ -36,18 +32,17 @@ public class AuthorizationService {
     return authClientImpl.getUserInfo(accessToken);
   }
 
-  public AccessTokenDTO postToken(String idToken) {
+  public AccessToken postToken(String idToken) {
     log.info("Posting token for validation");
 
-    return accessTokenDTOMapper.toDTO(
-      authClientImpl.postToken(
+    return authClientImpl.postToken(
         CLIENT_ID,
         GRANT_TYPE,
         SCOPE,
         idToken,
         subjectIssuer,
         SUBJECT_TOKEN_TYPE,
-        null));
+        null);
   }
 
   public void validateAdminRole(Long organizationId, UserInfo loggedUser) {
