+++ src/test/java/it/gov/pagopa/pu/bff/mapper/taxonomy/TaxonomyServiceTypeCodeMapperTest.java
@@ -1,18 +1,17 @@
-package it.gov.pagopa.pu.bff.mapper;
+package it.gov.pagopa.pu.bff.mapper.taxonomy;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import it.gov.pagopa.pu.bff.dto.generated.TaxonomyCollectionReasonDTO;
+import it.gov.pagopa.pu.bff.dto.generated.TaxonomyServiceTypeCodeDTO;
 import it.gov.pagopa.pu.bff.util.TestUtils;
 import org.junit.jupiter.api.Test;
 
-class TaxonomyCollectionReasonMapperTest {
-
-  private final TaxonomyCollectionReasonMapper mapper = new TaxonomyCollectionReasonMapper();
+class TaxonomyServiceTypeCodeMapperTest {
+  private final TaxonomyServiceTypeCodeMapper mapper = new TaxonomyServiceTypeCodeMapper();
 
   @Test
   void testMap() {
-    it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyCollectionReasonDTO input = new it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyCollectionReasonDTO();
+    it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyServiceTypeCodeDTO input = new it.gov.pagopa.pu.p4pa_organization.dto.generated.TaxonomyServiceTypeCodeDTO();
     input.setOrganizationType("Type1");
     input.setOrganizationTypeDescription("Description1");
     input.setMacroAreaCode("Macro1");
@@ -21,9 +20,8 @@ class TaxonomyCollectionReasonMapperTest {
     input.setServiceTypeCode("ServiceCode1");
     input.setServiceType("Service1");
     input.setServiceTypeDescription("ServiceDescription1");
-    input.setCollectionReason("Reason1");
 
-    TaxonomyCollectionReasonDTO output = mapper.map(input);
+    TaxonomyServiceTypeCodeDTO output = mapper.map(input);
 
     assertEquals("Type1. Description1", output.getOrganizationTypeDescription());
     assertEquals("Type1", output.getOrganizationType());
@@ -33,8 +31,6 @@ class TaxonomyCollectionReasonMapperTest {
     assertEquals("ServiceCode1", output.getServiceTypeCode());
     assertEquals("ServiceCode1. Service1", output.getServiceType());
     assertEquals("ServiceDescription1", output.getServiceTypeDescription());
-    assertEquals("Reason1", output.getCollectionReason());
     TestUtils.checkNotNullFields(output);
   }
-
 }
